# src/main/resources/sales_and_user_management_api-documentation.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "TreadX API Documentation",
    "version": "1.0.0",
    "description": "API documentation for TreadX application"
  },
  "servers": [
    {
      "url": "/api/v1",
      "description": "API v1"
    }
  ],
  "paths": {
...
    "/roles": {
      "get": {
        "tags": ["Role Management"],
        "summary": "Get all roles",
        "description": "Retrieves a list of all roles in the system. Requires PLATFORM_ADMIN role.",
        "security": [
          {
            "bearerAuth": ["PLATFORM_ADMIN"]
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved all roles",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/RoleResponseDTO"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Access denied - insufficient permissions"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/roles/{id}": {
      "delete": {
        "tags": ["Role Management"],
        "summary": "Delete role",
        "description": "Deletes a role from the system. Requires PLATFORM_ADMIN role.",
        "security": [
          {
            "bearerAuth": ["PLATFORM_ADMIN"]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the role",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successfully deleted the role"
          },
          "404": {
            "description": "Role not found"
          },
          "403": {
            "description": "Access denied - insufficient permissions"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/roles/{id}/permissions": {
      "get": {
        "tags": ["Role Management"],
        "summary": "Get role permissions",
        "description": "Retrieves all permissions associated with a specific role. Requires PLATFORM_ADMIN role.",
        "security": [
          {
            "bearerAuth": ["PLATFORM_ADMIN"]
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "description": "ID of the role",
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successfully retrieved role permissions",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Role not found"
          },
          "403": {
            "description": "Access denied - insufficient permissions"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
  }
} 